+++ test/integration/install/test_install.py
@@ -19,7 +19,7 @@ def test_install_tt(tt_cmd, tmpdir):
         f.write('tt:\n  app:\n    bin_dir:\n    inc_dir:\n')
 
     # Install latest tt.
-    install_cmd = [tt_cmd, "--cfg", configPath, "install", "tt=0.1.0"]
+    install_cmd = [tt_cmd, "--cfg", configPath, "install", "tt"]
     instance_process = subprocess.Popen(
         install_cmd,
         cwd=tmpdir,
@@ -41,7 +41,41 @@ def test_install_tt(tt_cmd, tmpdir):
         text=True
     )
     start_output = installed_program_process.stdout.readline()
-    assert re.search(r"commit: ccd9c0c", start_output)
+    assert re.search(r"Tarantool CLI version \d+.\d+.\d+", start_output)
+
+
+@pytest.mark.slow
+def test_install_tt_specific_version(tt_cmd, tmpdir):
+
+    configPath = os.path.join(tmpdir, config_name)
+    # Create test config
+    with open(configPath, 'w') as f:
+        f.write('tt:\n  app:\n    bin_dir:\n    inc_dir:\n')
+
+    # Install latest tt.
+    install_cmd = [tt_cmd, "--cfg", configPath, "install", "tt", "1.0.0"]
+    instance_process = subprocess.Popen(
+        install_cmd,
+        cwd=tmpdir,
+        stderr=subprocess.STDOUT,
+        stdout=subprocess.PIPE,
+        text=True
+    )
+
+    # Check that the process shutdowned correctly.
+    instance_process_rc = instance_process.wait()
+    assert instance_process_rc == 0
+
+    installed_cmd = [tmpdir + "/bin/tt", "version"]
+    installed_program_process = subprocess.Popen(
+        installed_cmd,
+        cwd=tmpdir + "/bin",
+        stderr=subprocess.STDOUT,
+        stdout=subprocess.PIPE,
+        text=True
+    )
+    start_output = installed_program_process.stdout.readline()
+    assert re.search(r"Tarantool CLI version 1.0.0", start_output)
 
 
 @pytest.mark.slow
@@ -54,7 +88,7 @@ def test_install_tarantool(tt_cmd, tmpdir):
     tmpdir_without_config = tempfile.mkdtemp()
 
     # Install latest tarantool.
-    install_cmd = [tt_cmd, "--cfg", config_path, "install", "tarantool=2.10.4", "-f"]
+    install_cmd = [tt_cmd, "--cfg", config_path, "install", "-f", "tarantool", "2.10.4"]
     instance_process = subprocess.Popen(
         install_cmd,
         cwd=tmpdir_without_config,
@@ -78,6 +112,7 @@ def test_install_tarantool(tt_cmd, tmpdir):
     run_output = installed_program_process.stdout.readline()
     assert re.search(r"Tarantool", run_output)
     assert os.path.exists(os.path.join(tmpdir, "my_inc", "include", "tarantool"))
+    assert os.path.exists(os.path.join(tmpdir, "bin", "tarantool_2.10.4"))
 
 
 @pytest.mark.slow
@@ -93,7 +128,7 @@ def test_install_tarantool_in_docker(tt_cmd, tmpdir):
     tmpdir_without_config = tempfile.mkdtemp()
 
     # Install latest tarantool.
-    install_cmd = [tt_cmd, "--cfg", config_path, "install", "tarantool", "-f", "--use-docker"]
+    install_cmd = [tt_cmd, "--cfg", config_path, "install", "-f", "tarantool", "--use-docker"]
     tt_process = subprocess.Popen(
         install_cmd,
         cwd=tmpdir_without_config,
@@ -125,28 +160,3 @@ def test_install_tarantool_in_docker(tt_cmd, tmpdir):
     assert out == "GLIBC_2.27"
 
     assert os.path.exists(os.path.join(tmpdir, "my_inc", "include", "tarantool"))
-
-
-def test_install_tt_in_docker(tt_cmd, tmpdir):
-    if platform.system() == "Darwin":
-        pytest.skip("/set platform is unsupported")
-
-    config_path = os.path.join(tmpdir, config_name)
-    with open(config_path, "w") as f:
-        yaml.dump({"tt": {"app": {"bin_dir": "", "inc_dir": "./my_inc"}}}, f)
-
-    tmpdir_without_config = tempfile.mkdtemp()
-
-    install_cmd = [tt_cmd, "--cfg", config_path, "install", "tt", "--use-docker"]
-    tt_process = subprocess.Popen(
-        install_cmd,
-        cwd=tmpdir_without_config,
-        stderr=subprocess.STDOUT,
-        stdout=subprocess.PIPE,
-        text=True
-    )
-
-    tt_process_rc = tt_process.wait()
-    assert tt_process_rc == 1
-    output = tt_process.stdout.readline()
-    assert re.search(r"--use-docker can be used only for 'tarantool' program", output)